refactored the code

+++ src/wasm/console.rs
@@ -2,7 +2,7 @@ use anyhow::{anyhow, Result};
 use wasmtime::{Caller, Func, Store};
 
 pub fn print_num(store: &mut Store<()>) -> Func {
-    Func::wrap(store, |_: Caller<'_, ()>, num: i32| -> Result<()> {
+    Func::wrap(store, |_: Caller<'_, ()>, num: f64| -> Result<()> {
         println!("{num}");
         Ok(())
     })
